tutorial_implementation/tutorial34/requirements.txt
# Tutorial 34: Google Cloud Pub/Sub + Event-Driven Agents - Python Dependencies

# Core ADK framework
google-adk>=1.15.1

# Google Cloud Pub/Sub
google-cloud-pubsub>=2.23.0

# Google AI API
google-genai>=1.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
